[WIP] Implement image upload feature for POST /dogs
Works via Postman (set fields via "form-data" instead of "raw JSON",
with these fields: name, description, gender, hdbApproved, image - where
the image is a jpg file), BUT does not work via the frontend.

When submitting the form via the frontend, the form values seems to
include the `File` in the values, but the Document saved in the db does not
have any 'file' attribute - so something in the form appears to be
missing.